feat(worldgen): add isAreaClearForBuilding helper in layout_core.js and reuse in town_gen.js to enforce building margins.
